requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.12.3
python-multipart==0.0.9
orjson==3.10.7
numpy==2.1.2
matplotlib==3.9.2
sentence-transformers==2.7.0
openai==1.40.0
anthropic==0.36.0
tenacity==8.2.3
jsonschema==4.23.0
requests==2.32.3
PyYAML==6.0.2
